================================================== Resolving issue #a79062 with 549b9c43 on master ================================================== The current stashing mechanism is broken.
Currently when git-issue stashes, it does it unconditional of the
branch actually being dirty. If a non-dirty branch stash pops a
stash-stack, then it will apply the stash--this is bad.

Type: bug
Milestone: initial-release
Tag: release-blocker
Assign: justinvh@gmail.com

--------------------------------------------------
With commit 85c7fd4f on master
--------------------------------------------------
Fixes git stashing issue on dirty HEAD

- Previously a git stash would happen regardless. Now git-issues checks
if the branch is dirty and stashes accordingly.

Fixes: a79062

git-issue